bug fix - writing to protected memory in NtUtility.h, causing crash when click apply after run
git-tfs-id: [http://tfs.bumc.bu.edu:8080/tfs/BU-MICRO-MSI]$/Daphne/Daphne-axin3;C6229

+++ NativeDaphneLibrary/NtCollisionManager.cpp
@@ -225,7 +225,7 @@ namespace NativeDaphneLibrary
 	NtCellPair* NtCollisionManager::NewCellPair(long key, NtCell* _a, NtCell* _b)
 	{
 		NtCellPair* cellpair = NULL;
-		unsigned count = FreePairList.size();
+		unsigned count = (unsigned)FreePairList.size();
 		if (count > 0)
 		{
 			cellpair = FreePairList[count-1];
@@ -247,7 +247,7 @@ namespace NativeDaphneLibrary
 	void NtCollisionManager::balance()
 	{
 		
-		unsigned count = FreePairList.size();
+		unsigned count = (unsigned)FreePairList.size();
 		if (count == 0)
 		{
 			numPairs = NextFreeIndex;
@@ -255,7 +255,7 @@ namespace NativeDaphneLibrary
 		}
 
 		NtCellPair *pairArray = PairArrayStorage;
-		for (int i=0; i< count; i++)
+		for (int i=0; i< (int)count; i++)
 		{
 			NtCellPair* cellpair = FreePairList[i];
 			int index = (int)(cellpair - pairArray);
